requirements.txt
Django
django-celery-with-redis
setproctitle
